# frontend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Media Platform API",
    "description": "Queue-based media playback controller for homelab",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5199/"
    }
  ],
  "paths": {
...
    "/api/v1/player/heartbeat": {
      "post": {
        "tags": [
          "Player"
        ],
        "description": "Report player liveness heartbeat",
        "operationId": "PlayerHeartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/player/play": {
      "post": {
        "tags": [
          "Player"
        ],
        "description": "Start or resume playback",
        "operationId": "Play",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackStateResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/player/pause": {
      "post": {
        "tags": [
          "Player"
        ],
        "description": "Pause playback",
        "operationId": "Pause",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackStateResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/player/skip": {
      "post": {
        "tags": [
          "Player"
        ],
        "description": "Skip to the next track",
        "operationId": "Skip",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackStateResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/player/stop": {
      "post": {
        "tags": [
          "Player"
        ],
        "description": "Stop playback",
        "operationId": "Stop",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackStateResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/player/position": {
      "post": {
        "tags": [
          "Player"
        ],
        "description": "Report current playback position",
        "operationId": "ReportPosition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportPositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/player/error": {
      "post": {
        "tags": [
          "Player"
        ],
        "description": "Report a playback error",
        "operationId": "ReportError",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportErrorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackStateResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "HeartbeatRequest": {
        "required": [
          "playerId",
          "state"
        ],
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "position": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double",
            "default": 0
          },
          "videoId": {
            "type": [
              "null",
              "string"
            ]
          },
          "uptime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64",
            "default": 0
          },
          "version": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PlaybackStateResponse": {
        "required": [
          "state",
          "currentItem",
          "startedAt",
          "positionSeconds",
          "retryCount",
          "lastError"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          },
          "currentItem": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/QueueItemResponse"
              }
            ]
          },
          "startedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "positionSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "retryCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastError": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "QueueItemResponse": {
        "required": [
          "id",
          "url",
          "title",
          "status",
          "addedAt",
          "startAtSeconds"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startAtSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "addedByUserId": {
            "type": [
              "null",
              "string"
            ]
          },
          "addedByName": {
            "type": [
              "null",
              "string"
            ]
          },
          "channel": {
            "type": [
              "null",
              "string"
            ]
          },
          "durationSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "thumbnailUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReportErrorRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "ReportPositionRequest": {
        "required": [
          "positionSeconds"
        ],
        "type": "object",
        "properties": {
          "positionSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
    }
  },
...
}